+++ src/install.rs
@@ -5,6 +5,7 @@ use std::{
     fmt::Write as _,
     fs,
     io::{BufRead, BufReader, BufWriter},
+    ops::{BitAnd, BitOr, BitOrAssign},
     path::Path,
     path::PathBuf,
 };
@@ -40,6 +41,9 @@ pub fn install_config(config_file: Option<&Path>, config_type: GitConfigType) ->
     let mut file = if file_path.is_file() {
         gix_config::File::from_path_no_includes(file_path.clone(), source)?
     } else {
+        if let Some(parent) = file_path.parent() {
+            fs::create_dir_all(parent)?;
+        }
         gix_config::File::new(gix_config::file::Metadata::from(source))
     };
 
@@ -84,7 +88,6 @@ fn resolve_config_file(
         Ok(config_file.to_path_buf())
     } else {
         let source: Source = config_type.into();
-        let cur_dir = std::env::current_dir()?;
         #[allow(clippy::unwrap_used)]
         let file_path = match config_type {
             GitConfigType::Global | GitConfigType::System => source
@@ -93,10 +96,7 @@ fn resolve_config_file(
                 .unwrap()
                 .to_owned(),
             GitConfigType::Local => {
-                let dotgit = gix_discover::upwards(&cur_dir)?
-                    .0
-                    .into_repository_and_work_tree_directories()
-                    .0;
+                let dotgit = get_git_repo_and_work_tree()?.0;
                 dotgit.join(source.storage_location(&mut gix_path::env::var).unwrap())
             }
         };
@@ -136,6 +136,25 @@ fn resolve_attribute_file(
             }
         };
         Ok(file_path)
+    }
+}
+
+fn get_git_repo_and_work_tree() -> Result<(PathBuf, Option<PathBuf>), Error> {
+    let cur_dir = std::env::current_dir()?;
+    Ok(gix_discover::upwards(&cur_dir)?
+        .0
+        .into_repository_and_work_tree_directories())
+}
+
+fn get_default_attribute_file() -> Result<Option<PathBuf>, Error> {
+    let Some(work_dir) = get_git_repo_and_work_tree()?.1 else {
+        return Ok(None);
+    };
+    let attribute_file = work_dir.join(".gitattributes");
+    if attribute_file.is_file() {
+        Ok(Some(attribute_file))
+    } else {
+        Ok(None)
     }
 }
 
@@ -184,6 +203,9 @@ pub fn install_attributes(
         }
     } else {
         println!("Writing to {}", file_path.display());
+        if let Some(parent) = file_path.parent() {
+            fs::create_dir_all(parent)?;
+        }
         let mut writer = fs::File::create(file_path)?;
 
         for line in ATTRIBUTE_LINES {
@@ -205,6 +227,9 @@ pub fn uninstall_attributes(
         let mut to_write = false;
         for line in f.lines() {
             let mut line = line?;
+            if line.is_empty() {
+                continue;
+            }
             // let (kind, x, _) = gix_attributes::parse(line.as_bytes()).next().unwrap()?;
             let (kind, x, _) = match gix_attributes::parse(line.as_bytes()).next() {
                 None => bail!("No pattern found in line"),
@@ -282,5 +307,205 @@ pub fn uninstall_config(
 
     Ok(())
 }
+
+#[derive(Debug, Clone, PartialEq, Eq, Default)]
+struct InstallToolStatus {
+    pub diff: bool,
+    pub filter: bool,
+}
+#[derive(Debug, Clone, PartialEq, Eq, Default)]
+struct InstallStatus {
+    pub nbstripout: InstallToolStatus,
+    pub nbwipers: InstallToolStatus,
+}
+
+impl InstallToolStatus {
+    fn is_installed(&self) -> bool {
+        self.diff && self.filter
+    }
+}
+// impl From<InstallToolStatus> for bool {
+//     fn from(value: InstallToolStatus) -> Self {
+//         value.diff & value.filter
+//     }
+// }
+// impl From<InstallStatus> for bool {
+//     fn from(value: InstallStatus) -> Self {
+//         (value.nbstripout | value.nbwipers).into()
+//     }
+// }
+
+impl BitAnd for InstallToolStatus {
+    type Output = InstallToolStatus;
+    fn bitand(self, rhs: Self) -> Self::Output {
+        InstallToolStatus {
+            diff: self.diff & rhs.diff,
+            filter: self.filter & rhs.filter,
+        }
+    }
+}
+
+impl BitOr for InstallToolStatus {
+    type Output = InstallToolStatus;
+    fn bitor(self, rhs: Self) -> Self::Output {
+        InstallToolStatus {
+            diff: self.diff | rhs.diff,
+            filter: self.filter | rhs.filter,
+        }
+    }
+}
+
+impl BitOr for InstallStatus {
+    type Output = InstallStatus;
+    fn bitor(self, rhs: Self) -> Self::Output {
+        InstallStatus {
+            nbstripout: self.nbstripout | rhs.nbstripout,
+            nbwipers: self.nbwipers | rhs.nbwipers,
+        }
+    }
+}
+impl BitAnd for InstallStatus {
+    type Output = InstallStatus;
+    fn bitand(self, rhs: Self) -> Self::Output {
+        InstallStatus {
+            nbstripout: self.nbstripout & rhs.nbstripout,
+            nbwipers: self.nbwipers & rhs.nbwipers,
+        }
+    }
+}
+
+impl BitOrAssign for InstallToolStatus {
+    fn bitor_assign(&mut self, rhs: Self) {
+        self.diff |= rhs.diff;
+        self.filter |= rhs.filter;
+    }
+}
+impl BitOrAssign for InstallStatus {
+    fn bitor_assign(&mut self, rhs: Self) {
+        self.nbstripout |= rhs.nbstripout;
+        self.nbwipers |= rhs.nbwipers;
+    }
+}
+
+fn check_attribute_file(attr_file_path: &Path) -> Result<InstallStatus, Error> {
+    if attr_file_path.is_file() {
+        let mut status = InstallStatus::default();
+
+        let bytes = fs::read(attr_file_path)?;
+        let lines = gix_attributes::parse(&bytes);
+        for line in lines {
+            let (kind, assignments, _) = line?;
+
+            if let Kind::Pattern(patt) = kind {
+                if patt.to_string() == "*.ipynb" {
+                    for assignment in assignments {
+                        let assignment = assignment?;
+                        if let StateRef::Value(s) = assignment.state {
+                            if s.as_bstr() == "nbwipers" {
+                                status.nbwipers.filter |= assignment.name.as_str() == "filter";
+                                status.nbwipers.diff |= assignment.name.as_str() == "diff";
+                            }
+                            if s.as_bstr() == "ipynb" {
+                                status.nbstripout.diff |= assignment.name.as_str() == "diff";
+                            }
+                            if s.as_bstr() == "nbstripout" {
+                                status.nbstripout.filter |= assignment.name.as_str() == "filter";
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Ok(status)
+    } else {
+        Ok(InstallStatus::default())
+    }
+}
+
+fn check_install_attr_files(config_types: &[GitConfigType]) -> Result<InstallStatus, Error> {
+    let attribute_file = get_default_attribute_file().ok().flatten();
+    let mut attr_install_status = match attribute_file {
+        Some(ref attr) => check_attribute_file(attr)?,
+        None => InstallStatus::default(),
+    };
+
+    for config_type in config_types {
+        let attr_file_path = resolve_attribute_file(*config_type, None)?;
+
+        attr_install_status |= check_attribute_file(&attr_file_path)?;
+    }
+    Ok(attr_install_status)
+}
+
+fn check_config_sections(
+    filter_section: &Result<&gix_config::file::Section, gix_config::lookup::existing::Error>,
+    diff_section: &Result<&gix_config::file::Section, gix_config::lookup::existing::Error>,
+) -> InstallToolStatus {
+    InstallToolStatus {
+        diff: diff_section
+            .as_ref()
+            .is_ok_and(|x| x.contains_key("textconv")),
+        filter: filter_section
+            .as_ref()
+            .is_ok_and(|x| x.contains_key("clean") && x.contains_key("smudge")),
+    }
+}
+
+fn check_install_config_file(config_file: &gix_config::File) -> InstallStatus {
+    let filter_section = config_file.section("filter", Some("nbwipers".into()));
+    let diff_section = config_file.section("diff", Some("nbwipers".into()));
+    let filter_section_nbstripout = config_file.section("filter", Some("nbstripout".into()));
+    let diff_section_nbstripout = config_file.section("diff", Some("ipynb".into()));
+
+    InstallStatus {
+        nbstripout: check_config_sections(&filter_section_nbstripout, &diff_section_nbstripout),
+        nbwipers: check_config_sections(&filter_section, &diff_section),
+    }
+}
+
+fn combine_install_status(
+    attr_install_status: InstallStatus,
+    config_install_status: InstallStatus,
+) -> Result<(), Error> {
+    let overall_status = attr_install_status & config_install_status;
+    let mut installed = false;
+    if overall_status.nbstripout.is_installed() {
+        installed = true;
+        println!("nbstripout is installed");
+    }
+    if overall_status.nbwipers.is_installed() {
+        installed = true;
+        println!("nbwipers is installed");
+    }
+    if installed {
+        Ok(())
+    } else {
+        bail!("Neither nbstripout nor nbwipers are installed.")
+    }
+}
+pub fn check_install_some_type(config_type: GitConfigType) -> Result<(), Error> {
+    let attr_install_status = check_install_attr_files(&[config_type])?;
+
+    let file_path = resolve_config_file(None, config_type)?;
+    let config_file =
+        gix_config::File::from_path_no_includes(file_path.clone(), config_type.into())?;
+    let config_install_status = check_install_config_file(&config_file);
+
+    combine_install_status(attr_install_status, config_install_status)
+}
+pub fn check_install_none_type() -> Result<(), Error> {
+    let config_types = vec![
+        GitConfigType::Local,
+        GitConfigType::Global,
+        GitConfigType::System,
+    ];
+
+    let attr_install_status = check_install_attr_files(&config_types)?;
+    let config_file = gix_config::File::from_git_dir(get_git_repo_and_work_tree()?.0)?;
+    let config_install_status = check_install_config_file(&config_file);
+
+    combine_install_status(attr_install_status, config_install_status)
+}
+
 #[cfg(test)]
 mod test {}